context: move linked nodes when assigning a driver
When a node that requires a driver is moved to a driver, collect all
linked nodes and move them to the driver as well.

This fixes the case where a single jack node (zita-n2j) is linked to
paplay. The zita-n2j node is moved to the dummy driver but paplay is not
and so nothing happens before this patch.