pulse-server: stream: remove `done` flag
When the `done` flag was first added in
9f9be7d7f2ee2fa464953e7a25a85bac18792241, it was
actually needed because cleanup was implemented
using a per-client eventfd which was signalled when
something related to the particular client needed
to be freed. The function that ran, then, checked
each stream's `done` flag, and freed them as necessary.

However, since c70a5de5266b878cb62cd0ac84fcf7a717146018,
the stream cleanup is done using a work queue, and as
a consequence, the `done` flag is no longer needed.