impl-client: add workaround for old pipewire-media-session compatibility
pipewire-media-session purposefully makes one of its cores to lag the
other, and then uses it to bind ids it obtained from the faster core.
This no longer works with the registry generation number checks.

It's possible to fix in p-m-s, but we can also add a specific workaround
for it.

This workaround is supposed to be eventually removed.  Workarounds for
other apps should not be added.